fix: make LoginScreen wait for API results to switch screens
